refactor: AccessToken 명 SocialToken 으로 변경 및 Auth 관련 패키지 분리
