+++ backend/baton/src/main/java/touch/baton/config/ArgumentResolverConfig.java
@@ -3,8 +3,8 @@ package touch.baton.config;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import touch.baton.domain.oauth.controller.AuthRunnerPrincipalArgumentResolver;
-import touch.baton.domain.oauth.controller.AuthSupporterPrincipalArgumentResolver;
+import touch.baton.domain.oauth.controller.resolver.AuthRunnerPrincipalArgumentResolver;
+import touch.baton.domain.oauth.controller.resolver.AuthSupporterPrincipalArgumentResolver;
 
 import java.util.List;
