+++ backend/baton/src/main/java/touch/baton/domain/oauth/controller/resolver/AuthRunnerPrincipalArgumentResolver.java
@@ -1,4 +1,4 @@
-package touch.baton.domain.oauth.controller;
+package touch.baton.domain.oauth.controller.resolver;
 
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
