+++ backend/baton/src/test/java/touch/baton/document/oauth/github/GithubOauthApiTest.java
@@ -57,7 +57,7 @@ class GithubOauthApiTest extends RestdocsConfig {
                 .andDo(print());
     }
 
-    @DisplayName("Github 소셜 로그인을 위해 AuthCode 를 받아 AccessToken 으로 교환하여 Github 프로필 정보를 찾아오고 미가입 사용자일 경우 자동으로 회원가입을 진행하고 JWT 로 변환하여 클라이언트에게 넘겨준다.")
+    @DisplayName("Github 소셜 로그인을 위해 AuthCode 를 받아 SocialToken 으로 교환하여 Github 프로필 정보를 찾아오고 미가입 사용자일 경우 자동으로 회원가입을 진행하고 JWT 로 변환하여 클라이언트에게 넘겨준다.")
     @Test
     void github_login() throws Exception {
         // given & when
